get rid of converter

+++ app/src/main/java/com/example/numberproject/data/local/entity/Number.kt
@@ -8,6 +8,6 @@ import java.math.BigInteger
 data class Number (
     @PrimaryKey(autoGenerate = true)
     val id : Int = 0,
-    val number : BigInteger,
+    val number : Long,
     val fact : String?,
 )
